{{cookiecutter.project_slug}}/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
sqlmodel = "^0.0.8"
uvicorn = "^0.21.1"
alembic = "^1.10.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tenacity = "^8.2.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
sqlalchemy = {version = ">=1.4.41,<1.5.0", extras = ["asyncio"]}
psycopg2-binary = "^2.9.6"
asyncpg = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
emails = "^0.6"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.2.2"
autoflake = "^2.0.2"
flake8 = "^6.0.0"

[tool.isort]
profile = "black"
line_length = 88
overwrite_in_place = true
skip = [".gitignore"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
warn_unused_ignores = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
